requirements.txt
flask==2.3.2
python-telegram-bot>=21.0.0
firebase-admin==6.2.0
httpx==0.27.0